Reports are in: 7.92 micro-seconds
This is honestly far lower than expected... but the ring scheduler is
super basic, and ruthlessly efficient.  Something to be proud of, I
guess...

Does anyone even read this sort of thing?  Well... if you are... thanks
for reading, and I hope you have a nice day!

I guess if you want a round robin scheduler implemented in a
linked-list-ring thing... then feel free to bother me on github, or
email at CptSpaceToaster@gmail.com